feat(106927): :sparkles: tipifica services
